After discussion with Jane and David on 10th April, did the following viral term edits:
1. Merged 
adhesion to symbiont ; GO:0051856
into:
adhesion to host ; GO:0044406
The adhesion is reciprocal. The viral child terms are meant for annotation to both viruses and hosts, but previously this didn't make sense with 'adhesion to host' as a parent. Therefore combined the two terms. NB: GO:51856 has no child terms.

2. Renamed the merged GO:0044406 term from:
adhesion to host ; GO:0044406
to:
adhesion of symbiont to host ; GO:0044406
Updated definition and synonym scope accordingly (to 'The attachment of a symbiont to its host'). Added comment that term can be used to annotate both host and symbiont GPs.

3. Removed parents from newly merged 'adhesion of symbiont to host ; GO:0044406' term:
- is_a: interaction with host (GO:0051701)
- is_a: interaction with symbiont (GO:0051702)

4. Added 'virion ; GO:0019012' to the virusGOslim.