Hi Stan
I'm afraid the correct process for adding terms that involve chemicals is to first add the chemical to CHEBI, and then make the GO term in TG template. That way the term gets all the correct logical definitions and placement in the graph.

I've added a ticket for 'elemental metal' to CHEBI:

https://sourceforge.net/p/chebi/curator-requests/2358/

When that term is in, you can go ahead and add your term in TG template.

Will obsolete this one. 
OBSOLETE GO:1990419 response to elemental metal